Refactor UI FastAPI app to unify chat and auth
